+++ test/Atata.WebDriverExtras.Tests/ElementNotMissingExceptionTests.cs
@@ -1,14 +1,14 @@
 ﻿namespace Atata.WebDriverExtras.Tests;
 
-public static class ElementExceptionFactoryTests
+public static class ElementNotMissingExceptionTests
 {
     [TestFixture]
-    public class CreateForNotMissing
+    public sealed class Create
     {
         [Test]
         public void WithNull()
         {
-            var result = ElementExceptionFactory.CreateForNotMissing(null!);
+            var result = ElementNotMissingException.Create(null!);
 
             string expectedMessage = new SearchFailureData().ToStringForElementNotMissing();
             Assert.That(result.Message, Is.EqualTo(expectedMessage));
@@ -17,13 +17,13 @@ public static class ElementExceptionFactoryTests
         [Test]
         public void WithData()
         {
-            SearchFailureData data = new SearchFailureData
+            SearchFailureData data = new()
             {
                 By = By.XPath(".//a"),
                 SearchOptions = SearchOptions.Visible()
             };
 
-            var result = ElementExceptionFactory.CreateForNotMissing(data);
+            var result = ElementNotMissingException.Create(data);
 
             string expectedMessage = data.ToStringForElementNotMissing();
             Assert.That(result.Message, Is.EqualTo(expectedMessage));